Embedded code in the Teensyduino
This code enables the connection between the teensyduino and the app on the mobile via bluetooth where two movements have been converted to speech:

1. For change in 'value' of flex sensor, the app will say 'DRINK WATER'.

2. For change in the 'event.orientation.y' of adafruit, the app will say 'HELP ME'.

NOTE: Flex sensor and the adafruit measures the finger and the axis movement respectively and are fabricated on the glove. Like wise many commands can be given for different values generated by the hand movements.